persistant storage

+++ 690Project/Project/FileSaver.cs
@@ -7,7 +7,9 @@ public class FileSaver
     string fileName;
     public FileSaver(string fileName) {
         this.fileName = fileName;
+        if(!File.Exists(this.fileName)){
         File.Create(this.fileName).Close();
+        }
     }
 
     public void AppendLine(string line)
